requirements.txt
Flask>=3.1.0
jmcomic>=2.5.32
Pillow>=11.1.0
PyMuPDF
waitress>=3.0.2
watchdog>=6.0.0
